moved account stuff to account.py; changed s3 functions to one image per call
