+++ api/routers/auth.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from fastapi import APIRouter, HTTPException, status, Depends, Header, Response, Request
+from fastapi import APIRouter, HTTPException, status, Depends, Response, Request
 import bcrypt
 import jwt
 import datetime
@@ -14,7 +14,7 @@ from google.oauth2 import id_token
 load_dotenv()
 
 from api.database import get_db
-from api.models import Users, VerificationCodes, Listings, Messages
+from api.models import Users, VerificationCodes
 from api.services.email_service import get_email_service
 from sqlalchemy.orm import Session
 
@@ -51,12 +51,6 @@ class ResendCode(BaseModel):
 class GoogleAuth(BaseModel):
     idToken: str
     profile: dict
-
-
-class UpdateProfile(BaseModel):
-    firstName: str
-    lastName: str
-
 
 def hash_password(password: str) -> str:
     """Hash password using bcrypt"""
@@ -333,112 +327,6 @@ def resend_verification(resend: ResendCode, db: Session = Depends(get_db)):
 
     return {
         "message": "Verification code sent successfully"
-    }
-
-
-@router.delete('/delete-account')
-def delete_account(token_data: dict = Depends(verify_jwt_token), db: Session = Depends(get_db)):
-    """Delete user account and all associated data"""
-    user_id = token_data['uuid']
-
-    # Import S3 service here to avoid circular imports
-    s3_service = None
-    try:
-        from api.services.s3_service import get_s3_service
-        s3_service = get_s3_service()
-        s3_available = True
-    except ImportError:
-        s3_available = False
-        print("S3 service not available, skipping image cleanup")
-
-    # Find the user
-    user = db.query(Users).filter(Users.id == user_id).first()
-
-    if not user:
-        raise HTTPException(
-            status_code = status.HTTP_404_NOT_FOUND,
-            detail = "User not found"
-        )
-
-    # Get all listing images to delete from S3
-    all_image_urls = []
-    if s3_available:
-        user_listings = db.query(Listings).filter(Listings.seller_id == user_id).all()
-        
-        for listing in user_listings:
-            if listing.images and isinstance(listing.images, list):
-                # Filter out placeholder images, only delete S3 images
-                s3_images = [url for url in listing.images if
-                             not url.startswith('https://placebear.com')]
-                all_image_urls.extend(s3_images)
-
-    # Delete all user's listings from database (cascade will handle this)
-    db.query(Listings).filter(Listings.seller_id == user_id).delete()
-
-    # Delete all messages where user is sender OR receiver
-    db.query(Messages).filter(Messages.sender_id == user_id).delete()
-    db.query(Messages).filter(Messages.receiver_id == user_id).delete()
-    
-    # Delete the user account
-    db.delete(user)
-    db.commit()
-
-    # Delete images from S3 after database transaction is complete
-    if s3_available and all_image_urls:
-        try:
-            s3_service.delete_listing_images(all_image_urls)
-            print(f"Deleted {len(all_image_urls)} images from S3 for deleted account")
-        except Exception as e:
-            print(f"Warning: Failed to delete some S3 images for deleted account: {str(e)}")
-
-    return {"message": "Account successfully deleted"}
-
-
-@router.get('/profile')
-def get_profile(token_data: dict = Depends(verify_jwt_token), db: Session = Depends(get_db)):
-    """Get user profile information"""
-    user_id = token_data['uuid']
-    
-    user = db.query(Users).filter(Users.id == user_id).first()
-    
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found"
-        )
-    
-    return {
-        "firstName": user.fname,
-        "lastName": user.lname,
-        "email": user.email,
-        "isGoogleUser": bool(user.google_id)
-    }
-
-
-@router.put('/profile')
-def update_profile(profile_data: UpdateProfile, token_data: dict = Depends(verify_jwt_token), db: Session = Depends(get_db)):
-    """Update user profile information"""
-    user_id = token_data['uuid']
-    
-    user = db.query(Users).filter(Users.id == user_id).first()
-    
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found"
-        )
-    
-    # Update user information
-    user.fname = profile_data.firstName.strip()
-    user.lname = profile_data.lastName.strip()
-    db.commit()
-    
-    return {
-        "message": "Profile updated successfully",
-        "user": {
-            "id": str(user.id),
-            "email": user.email
-        }
     }
