+++ api/services/s3_service.py
@@ -26,55 +26,54 @@ class S3Service:
             aws_secret_access_key=aws_secret_key,
             region_name=aws_region
         )
+
+    def upload_profile_picture(self, file_content: bytes, file_extension: str, user_id: str) -> str:
+        """Upload profile picture to S3 and return the public URL"""
+
+        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
+        filename = f"profilepics/{user_id}/{timestamp}.{file_extension}"
+
+        return self.upload_image(file_content, filename, file_extension)
+
         
     def upload_listing_image(self, file_content: bytes, file_extension: str, listing_id: str) -> str:
         """
         Upload a listing image to S3 and return the public URL
         """
+        # Generate unique filename
+        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
+        unique_id = str(uuid.uuid4())[:8]
+        filename = f"listings/{listing_id}/{timestamp}_{unique_id}.{file_extension}"
+
+        return self.upload_image(file_content, filename, file_extension)
+
+    def upload_image(self, file_content: bytes, filename: str, file_extension: str) -> str:
         try:
-            # Generate unique filename
-            timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-            unique_id = str(uuid.uuid4())[:8]
-            filename = f"listings/{listing_id}/{timestamp}_{unique_id}.{file_extension}"
-            
             # Upload to S3
             self.s3_client.put_object(
-                Bucket=self.bucket_name,
-                Key=filename,
-                Body=file_content,
-                ContentType=f'image/{file_extension}',
-                ACL='public-read'
+                Bucket = self.bucket_name,
+                Key = filename,
+                Body = file_content,
+                ContentType = f'image/{file_extension}',
+                ACL = 'public-read'
             )
-            
+
             # Return public URL
             return f"https://{self.bucket_name}.s3.{os.getenv('AWS_REGION', 'us-east-1')}.amazonaws.com/{filename}"
-            
         except ClientError as e:
             raise Exception(f"Failed to upload image: {str(e)}")
     
-    def upload_multiple_images(self, files_data: List[tuple], listing_id: str) -> List[str]:
-        """
-        Upload multiple images and return list of URLs
-        files_data: List of tuples (file_content: bytes, file_extension: str)
-        """
-        urls = []
-        for file_content, file_extension in files_data:
-            url = self.upload_listing_image(file_content, file_extension, listing_id)
-            urls.append(url)
-        return urls
-    
-    def delete_listing_images(self, image_urls: List[str]) -> bool:
+    def delete_image(self, image_url: str) -> bool:
         """
         Delete images from S3 given their URLs
         """
         try:
-            for url in image_urls:
-                # Extract key from URL
-                key = url.split(f'{self.bucket_name}.s3')[1].split('.amazonaws.com/')[1]
-                self.s3_client.delete_object(Bucket=self.bucket_name, Key=key)
+            # Extract key from URL
+            key = image_url.split(f'{self.bucket_name}.s3')[1].split('.amazonaws.com/')[1]
+            self.s3_client.delete_object(Bucket=self.bucket_name, Key=key)
             return True
         except ClientError as e:
-            print(f"Error deleting images: {str(e)}")
+            print(f"Error deleting image: {str(e)}")
             return False
 
 # Create global instance with lazy initialization
